fix(water): carry EPANET options, per-tank event steps, float64 MinorLoss, wider D7 gradients
Milestone-4 fix wave, water part. N5: the EPANET reader carries DEMAND MODEL, MINIMUM/REQUIRED PRESSURE, PRESSURE EXPONENT, SPECIFIC GRAVITY and VISCOSITY on WaterOptions; build_water_model defaults its pressure-driven-demand arguments from them and threads density and viscosity into the Darcy-Weisbach path (refused by name on Hazen-Williams); other option lines are recorded, not dropped; D5 now reads its options from the file. N6: D7 differences roughness and the pump h0 as well as the sources, and pins that a single steady solve is structurally independent of the tank area (the channel is TankLevels.advance). N12: MinorLoss transition constant at the working dtype. N13: TankLevels.event_step is per tank. N15/FR-17: water_eps.py docstring says sequential, parents[1]. FR-13: _refuse_pump_overflow uses element_for. FR-15/16: D7/D8 measured figures refreshed.

+++ src/tellegen/apps/water/elements.py
@@ -297,7 +297,11 @@ class MinorLoss(Element):
         mask = dp.abs() < dpt
         dp_safe = torch.where(mask, torch.full_like(dp, dpt), dp.abs())
         sharp = torch.sign(dp) * torch.sqrt(dp_safe / m)
-        slope = torch.sqrt(torch.as_tensor(dpt) / m) / dpt
+        # `torch.as_tensor(dpt)` on a bare Python float casts with
+        # `torch.get_default_dtype()` (float32 in this repo, `elements/duct.py`'s
+        # documented gotcha), which would silently downcast a float64 element; `dtype=
+        # dp.dtype` keeps it in the caller's own precision, as `Headspace` does.
+        slope = torch.sqrt(torch.as_tensor(dpt, dtype=dp.dtype) / m) / dpt
         return torch.where(mask, slope * dp, sharp)
 
     def dflow(self, dp: Tensor, drivers=None) -> Tensor:
@@ -306,12 +310,12 @@ class MinorLoss(Element):
         mask = dp.abs() < dpt
         dp_safe = torch.where(mask, torch.full_like(dp, dpt), dp.abs())
         sharp = 0.5 / torch.sqrt(m * dp_safe)
-        slope = torch.sqrt(torch.as_tensor(dpt) / m) / dpt
+        slope = torch.sqrt(torch.as_tensor(dpt, dtype=dp.dtype) / m) / dpt
         return torch.where(mask, slope * torch.ones_like(dp), sharp)
 
     def linear_init(self, drivers=None) -> tuple[Tensor, Tensor]:
         m = self.coefficient()
         dpt = self.dp_transition
-        slope = torch.sqrt(torch.as_tensor(dpt) / m) / dpt
+        slope = torch.sqrt(torch.as_tensor(dpt, dtype=m.dtype) / m) / dpt
         zero = torch.zeros_like(slope)
         return zero, slope + zero
